Add contract interactor
