+++ packages/evm-consensus/source/consensus-interactor.ts
@@ -1,0 +1,98 @@
+import { inject, injectable, tagged } from "@mainsail/container";
+import { Contracts, Identifiers } from "@mainsail/contracts";
+import { ConsensusAbi } from "@mainsail/evm-contracts";
+import { BigNumber } from "@mainsail/utils";
+import { ethers } from "ethers";
+
+import { Identifiers as EvmConsensusIdentifiers } from "./identifiers.js";
+
+@injectable()
+export class ConsensusContractInteractor {
+	@inject(Identifiers.Application.Instance)
+	private readonly app!: Contracts.Kernel.Application;
+
+	@inject(Identifiers.Cryptography.Configuration)
+	private readonly configuration!: Contracts.Crypto.Configuration;
+
+	@inject(Identifiers.Evm.Instance)
+	@tagged("instance", "evm")
+	private readonly evm!: Contracts.Evm.Instance;
+
+	async getActiveValidators(): Promise<Contracts.State.ValidatorWallet[]> {
+		const consensusContractAddress = this.app.get<string>(EvmConsensusIdentifiers.Contracts.Addresses.Consensus);
+		const deployerAddress = this.app.get<string>(EvmConsensusIdentifiers.Internal.Addresses.Deployer);
+		const { evmSpec } = this.configuration.getMilestone();
+
+		const iface = new ethers.Interface(ConsensusAbi.abi);
+		const data = iface.encodeFunctionData("getTopValidators").slice(2);
+
+		const result = await this.evm.view({
+			caller: deployerAddress,
+			data: Buffer.from(data, "hex"),
+			recipient: consensusContractAddress,
+			specId: evmSpec,
+		});
+
+		if (!result.success) {
+			void this.app.terminate("getTopValidators failed");
+		}
+
+		const [validators] = iface.decodeFunctionResult("getTopValidators", result.output!);
+
+		const validatorWallets: Contracts.State.ValidatorWallet[] = [];
+		for (const [, validator] of validators.entries()) {
+			const [address, [votersCount, voteBalance, isResigned, blsPublicKey]] = validator;
+
+			const validatorWallet: Contracts.State.ValidatorWallet = {
+				address,
+				blsPublicKey: blsPublicKey.slice(2),
+				isResigned,
+				voteBalance: BigNumber.make(voteBalance),
+				votersCount: Number(votersCount),
+			};
+
+			validatorWallets.push(validatorWallet);
+		}
+
+		return validatorWallets;
+	}
+
+	async getAllValidators(): Promise<Contracts.State.ValidatorWallet[]> {
+		const consensusContractAddress = this.app.get<string>(EvmConsensusIdentifiers.Contracts.Addresses.Consensus);
+		const deployerAddress = this.app.get<string>(EvmConsensusIdentifiers.Internal.Addresses.Deployer);
+		const { evmSpec } = this.configuration.getMilestone();
+
+		const iface = new ethers.Interface(ConsensusAbi.abi);
+		const data = iface.encodeFunctionData("getAllValidators").slice(2);
+
+		const result = await this.evm.view({
+			caller: deployerAddress,
+			data: Buffer.from(data, "hex"),
+			recipient: consensusContractAddress,
+			specId: evmSpec,
+		});
+
+		if (!result.success) {
+			this.app.terminate("getAllValidators failed");
+		}
+
+		const [validators] = iface.decodeFunctionResult("getAllValidators", result.output!);
+
+		const validatorWallets: Contracts.State.ValidatorWallet[] = [];
+		for (const [, validator] of validators.entries()) {
+			const [address, [votersCount, voteBalance, isResigned, blsPublicKey]] = validator;
+
+			const validatorWallet: Contracts.State.ValidatorWallet = {
+				address,
+				blsPublicKey: blsPublicKey.slice(2),
+				isResigned,
+				voteBalance: BigNumber.make(voteBalance),
+				votersCount: Number(votersCount),
+			};
+
+			validatorWallets.push(validatorWallet);
+		}
+
+		return validatorWallets;
+	}
+}
