+++ packages/evm-consensus/source/validator-set.ts
@@ -1,22 +1,16 @@
-import { inject, injectable, tagged } from "@mainsail/container";
+import { inject, injectable } from "@mainsail/container";
 import { Contracts, Exceptions, Identifiers } from "@mainsail/contracts";
-import { ConsensusAbi } from "@mainsail/evm-contracts";
 import { Utils } from "@mainsail/kernel";
-import { ethers } from "ethers";
 
-import { Identifiers as EvmConsensusIdentifiers } from "./identifiers.js";
+import { ConsensusContractInteractor } from "./consensus-interactor.js";
 
 @injectable()
 export class ValidatorSet implements Contracts.ValidatorSet.Service {
-	@inject(Identifiers.Application.Instance)
-	private readonly app!: Contracts.Kernel.Application;
-
 	@inject(Identifiers.Cryptography.Configuration)
 	private readonly configuration!: Contracts.Crypto.Configuration;
 
-	@inject(Identifiers.Evm.Instance)
-	@tagged("instance", "evm")
-	private readonly evm!: Contracts.Evm.Instance;
+	@inject(Identifiers.Evm.Interactor.Consensus)
+	private readonly consensusContractInteractor!: ConsensusContractInteractor;
 
 	#topValidators: Contracts.State.ValidatorWallet[] = [];
 	#indexByAddress: Map<string, number> = new Map();
@@ -27,7 +21,7 @@ export class ValidatorSet implements Contracts.ValidatorSet.Service {
 	public async restore(): Promise<void> {
 		await this.#buildActiveValidators();
 
-		const validators = await this.#getAllValidators();
+		const validators = await this.consensusContractInteractor.getAllValidators();
 		this.#allValidators = new Map(validators.map((validator) => [validator.address, validator]));
 	}
 
@@ -73,7 +67,7 @@ export class ValidatorSet implements Contracts.ValidatorSet.Service {
 
 	async #buildActiveValidators(): Promise<void> {
 		const { activeValidators } = this.configuration.getMilestone();
-		const validators = await this.#getActiveValidators();
+		const validators = await this.consensusContractInteractor.getActiveValidators();
 		if (validators.length < activeValidators) {
 			throw new Exceptions.NotEnoughActiveValidatorsError(this.#topValidators.length, activeValidators);
 		}
@@ -85,7 +79,7 @@ export class ValidatorSet implements Contracts.ValidatorSet.Service {
 	async #calculateChangedValidators(): Promise<void> {
 		this.#dirtyValidators = [];
 
-		const validators = await this.#getAllValidators();
+		const validators = await this.consensusContractInteractor.getAllValidators();
 		for (const validator of validators) {
 			const currentValidator = this.#allValidators.get(validator.address);
 			if (
@@ -100,83 +94,5 @@ export class ValidatorSet implements Contracts.ValidatorSet.Service {
 		}
 
 		this.#allValidators = new Map(validators.map((validator) => [validator.address, validator]));
-	}
-
-	async #getActiveValidators(): Promise<Contracts.State.ValidatorWallet[]> {
-		const consensusContractAddress = this.app.get<string>(EvmConsensusIdentifiers.Contracts.Addresses.Consensus);
-		const deployerAddress = this.app.get<string>(EvmConsensusIdentifiers.Internal.Addresses.Deployer);
-		const { evmSpec } = this.configuration.getMilestone();
-
-		const iface = new ethers.Interface(ConsensusAbi.abi);
-		const data = iface.encodeFunctionData("getTopValidators").slice(2);
-
-		const result = await this.evm.view({
-			caller: deployerAddress,
-			data: Buffer.from(data, "hex"),
-			recipient: consensusContractAddress,
-			specId: evmSpec,
-		});
-
-		if (!result.success) {
-			this.app.terminate("getTopValidators failed");
-		}
-
-		const [validators] = iface.decodeFunctionResult("getTopValidators", result.output!);
-
-		const validatorWallets: Contracts.State.ValidatorWallet[] = [];
-		for (const [, validator] of validators.entries()) {
-			const [address, [votersCount, voteBalance, isResigned, blsPublicKey]] = validator;
-
-			const validatorWallet: Contracts.State.ValidatorWallet = {
-				address,
-				blsPublicKey: blsPublicKey.slice(2),
-				isResigned,
-				voteBalance: Utils.BigNumber.make(voteBalance),
-				votersCount: Number(votersCount),
-			};
-
-			validatorWallets.push(validatorWallet);
-		}
-
-		return validatorWallets;
-	}
-
-	async #getAllValidators(): Promise<Contracts.State.ValidatorWallet[]> {
-		const consensusContractAddress = this.app.get<string>(EvmConsensusIdentifiers.Contracts.Addresses.Consensus);
-		const deployerAddress = this.app.get<string>(EvmConsensusIdentifiers.Internal.Addresses.Deployer);
-		const { evmSpec } = this.configuration.getMilestone();
-
-		const iface = new ethers.Interface(ConsensusAbi.abi);
-		const data = iface.encodeFunctionData("getAllValidators").slice(2);
-
-		const result = await this.evm.view({
-			caller: deployerAddress,
-			data: Buffer.from(data, "hex"),
-			recipient: consensusContractAddress,
-			specId: evmSpec,
-		});
-
-		if (!result.success) {
-			this.app.terminate("getAllValidators failed");
-		}
-
-		const [validators] = iface.decodeFunctionResult("getAllValidators", result.output!);
-
-		const validatorWallets: Contracts.State.ValidatorWallet[] = [];
-		for (const [, validator] of validators.entries()) {
-			const [address, [votersCount, voteBalance, isResigned, blsPublicKey]] = validator;
-
-			const validatorWallet: Contracts.State.ValidatorWallet = {
-				address,
-				blsPublicKey: blsPublicKey.slice(2),
-				isResigned,
-				voteBalance: Utils.BigNumber.make(voteBalance),
-				votersCount: Number(votersCount),
-			};
-
-			validatorWallets.push(validatorWallet);
-		}
-
-		return validatorWallets;
 	}
 }
